profiles (see commit notes)
The intent with this is to expand the covfiguration to include
installation attributes according to the hostname of the machine. For
example, if `cat /etc/hostname` returns `ncase` or `x1`, we would
include the graphical environment. If it's not, assume it's a VPC
profile, and only install the desired packages needed for that.